Fix calculator to scale to long, and fix vessel class to take long as cooldown volume

+++ com-mmxlabs-lng-scheduler/com.mmxlabs.scheduler.optimiser/src/main/java/com/mmxlabs/scheduler/optimiser/Calculator.java
@@ -83,7 +83,7 @@ public final class Calculator {
 	 * @param value
 	 * @return scaled version of value
 	 */
-	public static int scale(final int value) {
+	public static long scale(final int value) {
 		return ScaleFactor * value;
 	}
