+++ com-mmxlabs-lng-scheduler/com.mmxlabs.scheduler.optimiser/src/main/java/com/mmxlabs/scheduler/optimiser/components/IVesselClass.java
@@ -167,5 +167,5 @@ public interface IVesselClass {
 	 * 
 	 * @return scaled M3 of LNG
 	 */
-	int getCooldownVolume();
+	long getCooldownVolume();
 }
